requirements.txt
runpod
pillow
rembg